requirements.txt
numpy
pyaudio
pygame
Pillow
